--- frontend/src/app/(auth)/sign_up/_components/templates/SignUpConfirmTemplate/index.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import { SignUpConfirmation } from '../../organisms/SignUpConfirmation';
-
-export const SignUpConfirmTemplate = () => {
-  return (
-    <>
-      <SignUpConfirmation />
-    </>
-  );
-};
