--- frontend/src/app/(auth)/sign_up/_components/templates/SignUpThanksTemplate/index.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import { SignUpThanks } from '../../organisms/SignUpThanks';
-
-export const SignUpThanksTemplate = () => {
-  return (
-    <>
-      <SignUpThanks />
-    </>
-  );
-};
