--- frontend/src/app/(auth)/sign_up/thanks/page.tsx
@@ -1,10 +1,0 @@
-import { SignUpThanksTemplate } from '../_components/templates/SignUpThanksTemplate';
-
-export default async function SignUpThanksPage() {
-  // TODO: /sign_up/confirm以外からの遷移を不可能にする
-  return (
-    <>
-      <SignUpThanksTemplate />
-    </>
-  );
-}
