+++ frontend/src/app/(auth)/sign_up/types/index.d.ts
@@ -1,0 +1,1 @@
+export type PhaseType = 'input' | 'confirmation' | 'thanks';
